Update TodoController
This updates the TodoController in the server so that the todos will be displayed.  Previously, the method on the client did not match correctly with the method in the server and this caused an issue. The sorting still needs to be addressed, this simply makes it so that the todos will appear on the client side todos-view.