Test Service component
This adds tests from the previous lab that test what is currently being done in with the TodoService.  Tests will need to be added for the new components, but these are just to make sure that what is currently here is working correctly.  Some slight adjustments were made to make the tests/code fit the current lab.